requirements/base.txt
SQLAlchemy-Utils==0.41.2
alembic==1.12.1
asyncpg==0.29.0
fastapi==0.115.0
httpx==0.27.0
passlib==1.7.4
pillow==10.3.0
psycopg2==2.9.9
pydantic-settings==2.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
python-stdnum==2.1
pytz==2024.1
requests==2.31.0
sqlalchemy==2.0.23
uvicorn==0.31.0
